refactor model into base class
